services/producer-service/requirements.txt
fastapi
uvicorn
kafka-python
prometheus-client
